shopify-agent-adk/pyproject.toml
[project]
name = "ai-agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.3.1",
    "google-adk>=1.13.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.111.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

